feat: add complexity support for mutations (wip)
