+++ packages/go/cypher/test/cases.go
@@ -1,17 +1,17 @@
 // Copyright 2023 Specter Ops, Inc.
-// 
+//
 // Licensed under the Apache License, Version 2.0
 // you may not use this file except in compliance with the License.
 // You may obtain a copy of the License at
-// 
+//
 //     http://www.apache.org/licenses/LICENSE-2.0
-// 
+//
 // Unless required by applicable law or agreed to in writing, software
 // distributed under the License is distributed on an "AS IS" BASIS,
 // WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 // See the License for the specific language governing permissions and
 // limitations under the License.
-// 
+//
 // SPDX-License-Identifier: Apache-2.0
 
 package test
@@ -20,4 +20,5 @@ const (
 	PositiveTestCases  = "cases/positive_tests.json"
 	NegativeTestCases  = "cases/negative_tests.json"
 	FilteringTestCases = "cases/filtering_tests.json"
+	MutationTestCases  = "cases/mutation_tests.json"
 )
